# skills/fabius-cohors/references/agents/go/navallist/docs/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Navallist API
  description: >
    API for the Navallist application.
    This API manages collaborative checklist sessions (Trips) and agent interactions.
    It simplifies the standard ADK-go patterns into a flat "/session" structure.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local Development Server

paths:
  # --- Session Management ---
  /session:
    post:
      summary: Create Session (Trip)
      description: Starts a new checklist trip.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                  example: "captain_steve"
                app_name:
                  type: string
                  default: "navallist"
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'

  /session/{id}:
    get:
      summary: Get Session State
      description: Retrieves the full state of a specific session (polling endpoint).
      operationId: getSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The SHORTKEY session ID.
      responses:
        '200':
          description: The requested session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          description: Session not found.

  # --- Agent Interaction ---
  /session/{id}/event:
    post:
      summary: Send Event (Chat/Update)
      description: Sends a user message or interaction (text/image) to the agent.
      operationId: sendEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The SHORTKEY session ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  oneOf:
                    - type: string
                    - $ref: '#/components/schemas/GenaiContent'
      responses:
        '200':
          description: The agent's response events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'

components:
  schemas:
    # Core Models
    Session:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        state:
          type: object
          description: Key-value map of checklist items (e.g., engine_oil=true).
          additionalProperties: true
        last_update:
          type: integer
          format: int64

    Event:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, model, system]
        content:
          $ref: '#/components/schemas/GenaiContent'

    # Google GenAI Types (simplified)
    GenaiContent:
      type: object
      properties:
        role:
          type: string
          example: "user"
        parts:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              inlineData:
                type: object
                properties:
                  mimeType:
                    type: string
                  data:
                    type: string
                    format: byte